Make slider images responsive
I decided to change the implementation of the slider image so it's more responsive.
I swapped to a picture element with multiple sources so the image will be changed
if when the viewport falls below a breakpoint rather than than waiting for a slide
change to pick the appropriate image size.